Refactoring in progress.
